Modify register and extend flow

+++ contracts/aurans-manager/src/price.rs
@@ -4,7 +4,7 @@ use cosmwasm_std::{Coin, Deps, Uint128};
 
 use crate::{error::ContractError, state::PRICE_INFO};
 
-pub fn calc_price(deps: Deps, name: &str, years: &u8) -> Result<Coin, ContractError> {
+pub fn calc_price(deps: Deps, name: &str, years: &u64) -> Result<Coin, ContractError> {
     let name_len = name.len() as u8;
     let has_price = PRICE_INFO.has(deps.storage, name_len);
     if !has_price {
